Move target resolver package
