+++ pkg/utils/targetresolver/secret/secretrefresolver.go
@@ -10,8 +10,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	lsv1alpha1 "github.com/gardener/landscaper/apis/core/v1alpha1"
-	. "github.com/gardener/landscaper/pkg/deployer/lib/targetresolver"
 	lsutils "github.com/gardener/landscaper/pkg/utils"
+	. "github.com/gardener/landscaper/pkg/utils/targetresolver"
 )
 
 var _ TargetResolver = SecretRefResolver{}
